[PAGOPA-2609] fix: junit tests
