+++ pom.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <project
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd"
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd"
+        xmlns="http://maven.apache.org/POM/4.0.0"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
   <groupId>it.gov.pagopa</groupId>
   <artifactId>pagopa-fdr</artifactId>
@@ -181,6 +181,11 @@
       <version>4.1.76.Final</version>
       <classifier>osx-aarch_64</classifier>
     </dependency>
+    <dependency>
+      <groupId>io.quarkus</groupId>
+      <artifactId>quarkus-jacoco</artifactId>
+      <scope>test</scope>
+    </dependency>
 
 
   </dependencies>
@@ -239,6 +244,7 @@
           <systemPropertyVariables>
             <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
             <maven.home>${maven.home}</maven.home>
+            <quarkus.jacoco.excludes>org/openapi/quarkus/**,org/openapi/**</quarkus.jacoco.excludes>
           </systemPropertyVariables>
         </configuration>
       </plugin>
@@ -263,6 +269,41 @@
           </execution>
         </executions>
       </plugin>
+      <!--      <plugin>-->
+      <!--        <groupId>org.jacoco</groupId>-->
+      <!--        <artifactId>jacoco-maven-plugin</artifactId>-->
+      <!--        <version>0.8.7</version>-->
+      <!--        <configuration>-->
+      <!--          <excludes>-->
+      <!--            <exclude>org/openapi/quarkus/**</exclude>-->
+      <!--            <exclude>org/openapi/**</exclude>-->
+      <!--          </excludes>-->
+      <!--        </configuration>-->
+      <!--        <executions>-->
+      <!--          <execution>-->
+      <!--            <id>default-prepare-agent</id>-->
+      <!--            <goals>-->
+      <!--              <goal>prepare-agent</goal>-->
+      <!--            </goals>-->
+      <!--            <configuration>-->
+      <!--              <destFile>${project.build.directory}/coverage-reports/jacoco.exec</destFile>-->
+      <!--              <propertyName>surefireArgLine</propertyName>-->
+      <!--            </configuration>-->
+      <!--          </execution>-->
+      <!--          <execution>-->
+      <!--            <id>default-report</id>-->
+      <!--            <phase>test</phase>-->
+      <!--            <goals>-->
+      <!--              <goal>report</goal>-->
+      <!--            </goals>-->
+      <!--            <configuration>-->
+      <!--              <dataFile>${project.build.directory}/coverage-reports/jacoco.exec</dataFile>-->
+      <!--              <outputDirectory>${project.reporting.outputDirectory}/jacoco</outputDirectory>-->
+      <!--            </configuration>-->
+      <!--          </execution>-->
+      <!--        </executions>-->
+      <!--      </plugin>-->
+
     </plugins>
   </build>
   <profiles>
