+++ src/test/java/it/gov/pagopa/fdr/controller/internal/InternalOperationControllerIT.java
@@ -1,9 +1,9 @@
-package it.gov.pagopa.fdr.controller.organizations;
+package it.gov.pagopa.fdr.controller.internal;
 
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 
 @QuarkusIntegrationTest
-class OrganizationResourceIT extends OrganizationResourceTest {
+class InternalOperationControllerIT extends InternalOperationControllerTest {
   // Execute the same tests but in packaged mode.
 
 }
