--- src/test/java/it/gov/pagopa/fdr/controller/organizations/InternalOrganizationResourceIT.java
@@ -1,9 +1,0 @@
-package it.gov.pagopa.fdr.controller.organizations;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-class InternalOrganizationResourceIT extends InternalOrganizationResourceTest {
-  // Execute the same tests but in packaged mode.
-
-}
