+++ src/test/java/it/gov/pagopa/fdr/controller/organizations/OrganizationsControllerTest.java
@@ -2,10 +2,12 @@ package it.gov.pagopa.fdr.controller.organizations;
 
 import static io.restassured.RestAssured.given;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.EC_CODE;
+import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.EC_CODE_NOT_ENABLED;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.HEADER;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PSP_CODE;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PSP_CODE_2;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PSP_CODE_NOT_ENABLED;
+import static it.gov.pagopa.fdr.util.error.enums.AppErrorCodeMessageEnum.PSP_UNKNOWN;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.containsInAnyOrder;
@@ -22,6 +24,7 @@ import it.gov.pagopa.fdr.controller.model.error.ErrorResponse;
 import it.gov.pagopa.fdr.controller.model.flow.enums.ReportingFlowStatusEnum;
 import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsResponse;
 import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowResponse;
+import it.gov.pagopa.fdr.controller.model.payment.Payment;
 import it.gov.pagopa.fdr.controller.model.payment.enums.PaymentStatusEnum;
 import it.gov.pagopa.fdr.controller.model.payment.response.PaginatedPaymentsResponse;
 import it.gov.pagopa.fdr.test.util.AzuriteResource;
@@ -37,14 +40,14 @@ import org.junit.jupiter.api.Test;
 @QuarkusTestResource(MockServerTestResource.class)
 @QuarkusTestResource(MongoResource.class)
 @QuarkusTestResource(AzuriteResource.class)
-class InternalOrganizationResourceTest {
+class OrganizationsControllerTest {
 
   private static final String GET_ALL_PUBLISHED_FLOW_URL =
-      "/internal/organizations/%s/fdrs?" + ControllerConstants.PARAMETER_PSP + "=%s";
+      "/organizations/%s/fdrs?" + ControllerConstants.PARAMETER_PSP + "=%s";
   private static final String GET_REPORTING_FLOW_URL =
-      "/internal/organizations/%s/fdrs/%s/revisions/%s/psps/%s";
+      "/organizations/%s/fdrs/%s/revisions/%s/psps/%s";
   private static final String GET_REPORTING_FLOW_PAYMENTS_URL =
-      "/internal/organizations/%s/fdrs/%s/revisions/%s/psps/%s/payments";
+      "/organizations/%s/fdrs/%s/revisions/%s/psps/%s/payments";
 
   /** ############### getAllPublishedFlow ################ */
   @Test
@@ -92,7 +95,7 @@ class InternalOrganizationResourceTest {
   @DisplayName("ORGANIZATIONS - KO FDR-0708 - psp unknown")
   void testOrganization_getAllPublishedFlow_KO_FDR0708() {
     String pspUnknown = "PSP_UNKNOWN";
-    String url = GET_ALL_PUBLISHED_FLOW_URL.formatted(EC_CODE, pspUnknown);
+    String url = GET_ALL_PUBLISHED_FLOW_URL.formatted(EC_CODE, pspUnknown, 10, 10);
     ErrorResponse res =
         given()
             .header(HEADER)
@@ -102,9 +105,8 @@ class InternalOrganizationResourceTest {
             .statusCode(400)
             .extract()
             .as(ErrorResponse.class);
-    assertThat(res.getHttpStatusCode(), equalTo(400));
     assertThat(res.getHttpStatusDescription(), equalTo("Bad Request"));
-    assertThat(res.getAppErrorCode(), equalTo("FDR-0708"));
+    assertThat(res.getAppErrorCode(), equalTo(PSP_UNKNOWN.errorCode()));
     assertThat(res.getErrors(), hasSize(1));
     assertThat(
         res.getErrors(),
@@ -114,7 +116,7 @@ class InternalOrganizationResourceTest {
   @Test
   @DisplayName("ORGANIZATIONS - KO FDR-0709 - psp not enabled")
   void testOrganization_getAllPublishedFlow_KO_FDR0709() {
-    String url = GET_ALL_PUBLISHED_FLOW_URL.formatted(EC_CODE, PSP_CODE_NOT_ENABLED);
+    String url = GET_ALL_PUBLISHED_FLOW_URL.formatted(EC_CODE, PSP_CODE_NOT_ENABLED, 10, 10);
 
     ErrorResponse res =
         given()
@@ -133,13 +135,59 @@ class InternalOrganizationResourceTest {
                 "message", equalTo("Psp [%s] not enabled".formatted(PSP_CODE_NOT_ENABLED)))));
   }
 
+  @Test
+  @DisplayName("ORGANIZATIONS - KO FDR-0716 - creditor institution unknown")
+  void testOrganization_getAllPublishedFlow_KO_FDR0716() {
+    String ecUnknown = "EC_UNKNOWN";
+    String url = GET_ALL_PUBLISHED_FLOW_URL.formatted(ecUnknown, PSP_CODE, 10, 10);
+
+    ErrorResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(400)
+            .extract()
+            .as(ErrorResponse.class);
+    assertThat(res.getAppErrorCode(), equalTo(AppErrorCodeMessageEnum.EC_UNKNOWN.errorCode()));
+    assertThat(
+        res.getErrors(),
+        hasItem(
+            hasProperty(
+                "message", equalTo("Creditor institution [%s] unknown".formatted(ecUnknown)))));
+  }
+
+  @Test
+  @DisplayName("ORGANIZATIONS - KO FDR-0717 - creditor institution not enabled")
+  void testOrganization_getAllPublishedFlow_KO_FDR0717() {
+    String url = GET_ALL_PUBLISHED_FLOW_URL.formatted(EC_CODE_NOT_ENABLED, PSP_CODE, 10, 10);
+
+    ErrorResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(400)
+            .extract()
+            .as(ErrorResponse.class);
+    assertThat(res.getAppErrorCode(), equalTo(AppErrorCodeMessageEnum.EC_NOT_ENABLED.errorCode()));
+    assertThat(
+        res.getErrors(),
+        hasItem(
+            hasProperty(
+                "message",
+                equalTo("Creditor institution [%s] not enabled".formatted(EC_CODE_NOT_ENABLED)))));
+  }
+
   /** ################# getReportingFlow ############### */
   @Test
   @DisplayName("ORGANIZATIONS - OK - recupero di un reporting flow")
   void testOrganization_getReportingFlow_Ok() {
     String flowName = TestUtil.getDynamicFlowName();
     TestUtil.pspSunnyDay(flowName);
-    String url = GET_REPORTING_FLOW_URL.formatted(EC_CODE, flowName, 1, PSP_CODE);
+    String url = GET_REPORTING_FLOW_URL.formatted(EC_CODE, flowName, 1L, PSP_CODE);
     SingleFlowResponse res =
         given()
             .header(HEADER)
@@ -163,7 +211,7 @@ class InternalOrganizationResourceTest {
     TestUtil.pspSunnyDay(flowName);
     TestUtil.pspSunnyDay(flowName);
 
-    String url = GET_REPORTING_FLOW_URL.formatted(EC_CODE, flowName, 2, PSP_CODE);
+    String url = GET_REPORTING_FLOW_URL.formatted(EC_CODE, flowName, 2L, PSP_CODE);
     SingleFlowResponse res =
         given()
             .header(HEADER)
@@ -185,7 +233,7 @@ class InternalOrganizationResourceTest {
     TestUtil.pspSunnyDay(flowName);
 
     String flowNameWrong = TestUtil.getDynamicFlowName();
-    String url = GET_REPORTING_FLOW_URL.formatted(EC_CODE, flowNameWrong, 1, PSP_CODE);
+    String url = GET_REPORTING_FLOW_URL.formatted(EC_CODE, flowNameWrong, 1L, PSP_CODE);
 
     ErrorResponse res =
         given()
@@ -212,7 +260,8 @@ class InternalOrganizationResourceTest {
     String flowName = TestUtil.getDynamicFlowName();
     TestUtil.pspSunnyDay(flowName);
 
-    String url = GET_REPORTING_FLOW_PAYMENTS_URL.formatted(EC_CODE, flowName, 1, PSP_CODE);
+    String url = GET_REPORTING_FLOW_PAYMENTS_URL.formatted(EC_CODE, flowName, 1L, PSP_CODE);
+
     PaginatedPaymentsResponse res =
         given()
             .header(HEADER)
@@ -223,7 +272,7 @@ class InternalOrganizationResourceTest {
             .extract()
             .as(PaginatedPaymentsResponse.class);
     assertThat(res.getCount(), equalTo(5L));
-    List<String> expectedList =
+    List expectedList =
         List.of(
             PaymentStatusEnum.EXECUTED.name(),
             PaymentStatusEnum.REVOKED.name(),
@@ -237,53 +286,33 @@ class InternalOrganizationResourceTest {
         containsInAnyOrder(expectedList.toArray()));
   }
 
-  /** ################# changeReadFlag ############### */
-  //  @Test
-  //  @DisplayName("ORGANIZATIONS - OK - changeReadFlag")
-  //  void testOrganization_changeReadFlag_Ok() {
-  //    String flowName = TestUtil.getDynamicFlowName();
-  //    TestUtil.pspSunnyDay(flowName);
-  //
-  //    String url = CHANGE_READ_FLAG_URL.formatted(flowName, 1, PSP_CODE);
-  //
-  //    GenericResponse res =
-  //        given()
-  //            .header(HEADER)
-  //            .when()
-  //            .put(url)
-  //            .then()
-  //            .statusCode(200)
-  //            .extract()
-  //            .as(GenericResponse.class);
-  //    assertThat(res.getMessage(), equalTo(String.format("Fdr [%s] internal read", flowName)));
-  //  }
-  //
-  //  @Test
-  //  @DisplayName("ORGANIZATIONS - KO FDR-0701 - changeReadFlag reporting flow not found")
-  //  void testOrganization_changeReadFlag_KO_FDR0701() {
-  //    String flowName = TestUtil.getDynamicFlowName();
-  //    TestUtil.pspSunnyDay(flowName);
-  //    String flowNameWrong = TestUtil.getDynamicFlowName();
-  //
-  //    String url = CHANGE_READ_FLAG_URL.formatted(flowNameWrong, 1, PSP_CODE);
-  //
-  //    ErrorResponse res =
-  //        given()
-  //            .header(HEADER)
-  //            .when()
-  //            .put(url)
-  //            .then()
-  //            .statusCode(404)
-  //            .extract()
-  //            .as(ErrorResponse.class);
-  //    assertThat(
-  //        res.getAppErrorCode(),
-  //        equalTo(AppErrorCodeMessageEnum.REPORTING_FLOW_NOT_FOUND.errorCode()));
-  //    assertThat(
-  //        res.getErrors(),
-  //        hasItem(
-  //            hasProperty(
-  //                "message",
-  //                equalTo(String.format("Fdr [%s] not found", flowNameWrong)))));
-  //  }
+  @Test
+  @DisplayName(
+      "ORGANIZATIONS - OK - recupero dei payments di un flow pubblicato con paginazione custom")
+  void testOrganization_getReportingFlowPayments_pagination_Ok() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+
+    String url =
+        (GET_REPORTING_FLOW_PAYMENTS_URL + "?page=2&size=1")
+            .formatted(EC_CODE, flowName, 1L, PSP_CODE);
+    PaginatedPaymentsResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(200)
+            .extract()
+            .as(PaginatedPaymentsResponse.class);
+    List<Payment> data = res.getData();
+
+    assertThat(res.getMetadata().getPageSize(), equalTo(1));
+    assertThat(res.getMetadata().getPageNumber(), equalTo(2));
+    assertThat(res.getCount(), equalTo(5L));
+    assertThat(
+        data.stream().map(o -> o.getPayStatus().name()).toList(),
+        equalTo(List.of(PaymentStatusEnum.REVOKED.name())));
+    assertThat(data.stream().map(Payment::getIndex).toList(), equalTo(List.of(101L)));
+  }
 }
