+++ src/test/java/it/gov/pagopa/fdr/controller/psps/PspsControllerIT.java
@@ -3,7 +3,7 @@ package it.gov.pagopa.fdr.controller.psps;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 
 @QuarkusIntegrationTest
-class InternalPspResourceIT extends PspResourceTest {
+class PspsControllerIT extends PspsControllerTest {
   // Execute the same tests but in packaged mode.
 
 }
