+++ src/test/java/it/gov/pagopa/fdr/controller/psps/PspsControllerTest.java
@@ -1,6 +1,7 @@
 package it.gov.pagopa.fdr.controller.psps;
 
 import static io.restassured.RestAssured.given;
+import static io.smallrye.common.constraint.Assert.assertTrue;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.BROKER_CODE;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.BROKER_CODE_2;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.BROKER_CODE_NOT_ENABLED;
@@ -8,30 +9,47 @@ import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.CHANNEL_CODE;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.CHANNEL_CODE_NOT_ENABLED;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.EC_CODE;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.EC_CODE_NOT_ENABLED;
+import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.FLOWS_DELETE_URL;
+import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.FLOWS_PUBLISH_URL;
+import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.FLOWS_URL;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.HEADER;
+import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PAYMENTS_ADD_URL;
+import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PAYMENTS_DELETE_URL;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PSP_CODE;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PSP_CODE_2;
+import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PSP_CODE_3;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.PSP_CODE_NOT_ENABLED;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.REPORTING_FLOW_NAME_DATE_WRONG_FORMAT;
 import static it.gov.pagopa.fdr.test.util.AppConstantTestHelper.REPORTING_FLOW_NAME_PSP_WRONG_FORMAT;
 import static it.gov.pagopa.fdr.test.util.TestUtil.FLOW_TEMPLATE;
 import static it.gov.pagopa.fdr.test.util.TestUtil.PAYMENTS_ADD_TEMPLATE;
 import static it.gov.pagopa.fdr.test.util.TestUtil.PAYMENTS_ADD_TEMPLATE_2;
+import static it.gov.pagopa.fdr.util.error.enums.AppErrorCodeMessageEnum.PSP_UNKNOWN;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.hasProperty;
+import static org.hamcrest.Matchers.hasSize;
 
 import io.quarkiverse.mockserver.test.MockServerTestResource;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 import it.gov.pagopa.fdr.controller.model.common.response.GenericResponse;
 import it.gov.pagopa.fdr.controller.model.error.ErrorResponse;
+import it.gov.pagopa.fdr.controller.model.flow.enums.ReportingFlowStatusEnum;
 import it.gov.pagopa.fdr.controller.model.flow.enums.SenderTypeEnum;
+import it.gov.pagopa.fdr.controller.model.flow.response.PaginatedFlowsCreatedResponse;
+import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowResponse;
+import it.gov.pagopa.fdr.controller.model.payment.Payment;
+import it.gov.pagopa.fdr.controller.model.payment.enums.PaymentStatusEnum;
+import it.gov.pagopa.fdr.controller.model.payment.response.PaginatedPaymentsResponse;
 import it.gov.pagopa.fdr.test.util.AzuriteResource;
 import it.gov.pagopa.fdr.test.util.MongoResource;
 import it.gov.pagopa.fdr.test.util.TestUtil;
 import it.gov.pagopa.fdr.util.error.enums.AppErrorCodeMessageEnum;
+import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -39,13 +57,16 @@ import org.junit.jupiter.api.Test;
 @QuarkusTestResource(MockServerTestResource.class)
 @QuarkusTestResource(MongoResource.class)
 @QuarkusTestResource(AzuriteResource.class)
-class InternalPspResourceTest {
+class PspsControllerTest {
 
-  private static final String FLOWS_URL = "/internal/psps/%s/fdrs/%s";
-  private static final String FLOWS_PUBLISH_URL = "/internal/psps/%s/fdrs/%s/publish";
-  private static final String FLOWS_DELETE_URL = "/internal/psps/%s/fdrs/%s";
-  private static final String PAYMENTS_ADD_URL = "/internal/psps/%s/fdrs/%s/payments/add";
-  private static final String PAYMENTS_DELETE_URL = "/internal/psps/%s/fdrs/%s/payments/del";
+  private static final String GET_FDR_CREATED_URL = "/psps/%s/created/fdrs/%s/organizations/%s";
+  private static final String GET_PAYMENTS_FDR_PUBLISHED_URL =
+      "/psps/%s/published/fdrs/%s/revisions/%s/organizations/%s/payments";
+  private static final String GET_FDR_PUBLISHED_URL =
+      "/psps/%s/published/fdrs/%s/revisions/%s/organizations/%s";
+  private static final String GET_ALL_FDR_CREATED_URL = "/psps/%s/created";
+  private static final String GET_PAYMENTS_FDR_CREATED_URL =
+      "/psps/%s/created/fdrs/%s/organizations/%s/payments";
 
   protected static String PAYMENTS_SAME_INDEX_ADD_TEMPLATE =
       """
@@ -194,8 +215,7 @@ class InternalPspResourceTest {
             100,
             101,
             102,
-            103,
-            104
+            103
         ]
       }
       """;
@@ -769,7 +789,7 @@ class InternalPspResourceTest {
   }
 
   @Test
-  @DisplayName("PSPS - KO FDR-0703 - fdr wrong action delete payments")
+  @DisplayName("PSPS - KO FDR-0703 - reporting flow wrong action delete payments")
   void test_psp_payments_delete_KO_FDR0703() {
     String flowName = TestUtil.getDynamicFlowName();
     String urlSave = FLOWS_URL.formatted(PSP_CODE, flowName);
@@ -852,8 +872,8 @@ class InternalPspResourceTest {
                 "message",
                 equalTo(
                     String.format(
-                        "Fdr [%s] have sender.pspId [%s] but not match with query param" + " [%s]",
-                        flowName, pspNotMatch, PSP_CODE)))));
+                        "Fdr [%s] have sender.pspId [%s] but not match with query param [%s]",
+                        flowName, PSP_CODE, pspNotMatch)))));
   }
 
   @Test
@@ -1140,7 +1160,7 @@ class InternalPspResourceTest {
             .statusCode(400)
             .extract()
             .as(ErrorResponse.class);
-    assertThat(res.getAppErrorCode(), equalTo(AppErrorCodeMessageEnum.PSP_UNKNOWN.errorCode()));
+    assertThat(res.getAppErrorCode(), equalTo(PSP_UNKNOWN.errorCode()));
     assertThat(
         res.getErrors(),
         hasItem(hasProperty("message", equalTo(String.format("Psp [%s] unknown", pspUnknown)))));
@@ -1384,7 +1404,7 @@ class InternalPspResourceTest {
 
   @Test
   @DisplayName("PSPS - KO FDR-0716 - ec unknown")
-  void test_organizationId_KO_FDR0716() {
+  void test_ecId_KO_FDR0716() {
     String flowName = TestUtil.getDynamicFlowName();
     String ecUnknown = "EC_UNKNOWN";
     String url = FLOWS_URL.formatted(PSP_CODE, flowName);
@@ -1418,7 +1438,7 @@ class InternalPspResourceTest {
 
   @Test
   @DisplayName("PSPS - KO FDR-0717 - ec not enabled")
-  void test_organizationId_KO_FDR0717() {
+  void test_ecId_KO_FDR0717() {
     String flowName = TestUtil.getDynamicFlowName();
     String url = FLOWS_URL.formatted(PSP_CODE, flowName);
     String bodyFmt =
@@ -1541,8 +1561,7 @@ class InternalPspResourceTest {
             .as(ErrorResponse.class);
     assertThat(res.getAppErrorCode(), equalTo(AppErrorCodeMessageEnum.BAD_REQUEST.errorCode()));
     assertThat(res.getErrors(), hasItem(hasProperty("message", equalTo("non deve essere null"))));
-    assertThat(
-        res.getErrors(), hasItem(hasProperty("path", equalTo("internalCreate.createRequest.fdr"))));
+    assertThat(res.getErrors(), hasItem(hasProperty("path", equalTo("createEmptyFlow.request.fdr"))));
   }
 
   @Test
@@ -1702,5 +1721,324 @@ class InternalPspResourceTest {
     assertThat(
         res.getErrors(),
         hasItem(hasProperty("message", equalTo("Bad request. Json format not valid"))));
+  }
+
+  /** ############### getAllPublishedFlow ################ */
+  @Test
+  @DisplayName("PSPS - OK - getAllPublishedFlow")
+  void testOrganization_getAllPublishedFlow_Ok() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+    String url = GET_FDR_PUBLISHED_URL.formatted(PSP_CODE, flowName, 1, EC_CODE);
+    SingleFlowResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(200)
+            .extract()
+            .as(SingleFlowResponse.class);
+    assertThat(res.getTotPayments(), equalTo(5L));
+    assertThat(res.getStatus(), equalTo(ReportingFlowStatusEnum.PUBLISHED));
+  }
+
+  @Test
+  @DisplayName("ORGANIZATIONS - OK - getAllPublishedFlow no results")
+  void test_psp_getAllPublishedFlow_OkNoResults() {
+    String flowName = TestUtil.getDynamicFlowName();
+
+    String urlPspFlow = FLOWS_URL.formatted(PSP_CODE, flowName);
+    String bodyFmtPspFlow =
+        FLOW_TEMPLATE.formatted(
+            flowName,
+            SenderTypeEnum.LEGAL_PERSON.name(),
+            PSP_CODE,
+            BROKER_CODE,
+            CHANNEL_CODE,
+            EC_CODE);
+
+    GenericResponse resPspFlow =
+        given()
+            .body(bodyFmtPspFlow)
+            .header(HEADER)
+            .when()
+            .post(urlPspFlow)
+            .then()
+            .statusCode(201)
+            .extract()
+            .body()
+            .as(GenericResponse.class);
+    assertThat(resPspFlow.getMessage(), equalTo(String.format("Fdr [%s] saved", flowName)));
+
+    String url = GET_FDR_PUBLISHED_URL.formatted(PSP_CODE, flowName, 1, EC_CODE);
+    ErrorResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(404)
+            .extract()
+            .as(ErrorResponse.class);
+    assertThat(
+        res.getAppErrorCode(),
+        equalTo(AppErrorCodeMessageEnum.REPORTING_FLOW_NOT_FOUND.errorCode()));
+  }
+
+  @Test
+  @DisplayName("PSPS - KO FDR-0708 - psp unknown")
+  void test_psp_getAllPublishedFlow_KO_FDR0708() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+    String pspUnknown = "PSP_UNKNOWN";
+    String url = GET_FDR_PUBLISHED_URL.formatted(pspUnknown, flowName, 1, EC_CODE);
+    ErrorResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(400)
+            .extract()
+            .as(ErrorResponse.class);
+    assertThat(res.getHttpStatusDescription(), equalTo("Bad Request"));
+    assertThat(res.getAppErrorCode(), equalTo(PSP_UNKNOWN.errorCode()));
+    assertThat(res.getErrors(), hasSize(1));
+    assertThat(
+        res.getErrors(),
+        hasItem(hasProperty("message", equalTo(String.format("Psp [%s] unknown", pspUnknown)))));
+  }
+
+  @Test
+  @DisplayName("PSPS - KO FDR-0709 - psp not enabled")
+  void test_psp_getAllPublishedFlow_KO_FDR0709() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+    String url = GET_FDR_PUBLISHED_URL.formatted(PSP_CODE_NOT_ENABLED, flowName, 1, EC_CODE);
+
+    ErrorResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(400)
+            .extract()
+            .as(ErrorResponse.class);
+    assertThat(res.getAppErrorCode(), equalTo(AppErrorCodeMessageEnum.PSP_NOT_ENABLED.errorCode()));
+    assertThat(
+        res.getErrors(),
+        hasItem(
+            hasProperty(
+                "message", equalTo("Psp [%s] not enabled".formatted(PSP_CODE_NOT_ENABLED)))));
+  }
+
+  @Test
+  @DisplayName("PSPS - KO FDR-0716 - creditor institution unknown")
+  void test_psp_getAllPublishedFlow_KO_FDR0716() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+    String ecUnknown = "EC_UNKNOWN";
+    String url = GET_FDR_PUBLISHED_URL.formatted(PSP_CODE, flowName, 1, ecUnknown);
+
+    ErrorResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(400)
+            .extract()
+            .as(ErrorResponse.class);
+    assertThat(res.getAppErrorCode(), equalTo(AppErrorCodeMessageEnum.EC_UNKNOWN.errorCode()));
+    assertThat(
+        res.getErrors(),
+        hasItem(
+            hasProperty(
+                "message", equalTo("Creditor institution [%s] unknown".formatted(ecUnknown)))));
+  }
+
+  @Test
+  @DisplayName("PSPS - KO FDR-0717 - creditor institution not enabled")
+  void test_psp_getAllPublishedFlow_KO_FDR0717() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+    String url = GET_FDR_PUBLISHED_URL.formatted(PSP_CODE, flowName, 1, EC_CODE_NOT_ENABLED);
+
+    ErrorResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(400)
+            .extract()
+            .as(ErrorResponse.class);
+    assertThat(res.getAppErrorCode(), equalTo(AppErrorCodeMessageEnum.EC_NOT_ENABLED.errorCode()));
+    assertThat(
+        res.getErrors(),
+        hasItem(
+            hasProperty(
+                "message",
+                equalTo("Creditor institution [%s] not enabled".formatted(EC_CODE_NOT_ENABLED)))));
+  }
+
+  /** ################# getReportingFlow ############### */
+  @Test
+  @DisplayName("PSPS - OK - recupero di un reporting flow")
+  void test_psp_getReportingFlow_Ok() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+    String url = GET_FDR_PUBLISHED_URL.formatted(PSP_CODE, flowName, 1L, EC_CODE);
+    SingleFlowResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(200)
+            .extract()
+            .as(SingleFlowResponse.class);
+    assertThat(res.getFdr(), equalTo(flowName));
+    assertThat(res.getReceiver().getOrganizationId(), equalTo(EC_CODE));
+    assertThat(res.getSender().getPspId(), equalTo(PSP_CODE));
+    assertThat(res.getStatus(), equalTo(ReportingFlowStatusEnum.PUBLISHED));
+    assertThat(res.getComputedTotPayments(), equalTo(5L));
+  }
+
+  @Test
+  @DisplayName("PSPS - OK - recupero di un reporting flow pubblicato alla revision 2")
+  void test_psp_getReportingFlow_revision_2_OK() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+    TestUtil.pspSunnyDay(flowName);
+
+    String url = GET_FDR_PUBLISHED_URL.formatted(PSP_CODE, flowName, 2L, EC_CODE);
+    SingleFlowResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(200)
+            .extract()
+            .as(SingleFlowResponse.class);
+    assertThat(res.getFdr(), equalTo(flowName));
+    assertThat(res.getRevision(), equalTo(2L));
+    assertThat(res.getStatus(), equalTo(ReportingFlowStatusEnum.PUBLISHED));
+  }
+
+  @Test
+  @DisplayName("PSPS - OK - nessun flusso trovato in stato CREATED per uno specifico PSP")
+  void test_psp_getAllReportingFlowCreated_OK() {
+    String url = (GET_ALL_FDR_CREATED_URL + "?page=2&size=1").formatted(PSP_CODE_3);
+
+    PaginatedFlowsCreatedResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(200)
+            .extract()
+            .as(PaginatedFlowsCreatedResponse.class);
+    assertThat(res.getCount(), equalTo(0L));
+  }
+
+  /** ################# getReportingFlowPayments ############### */
+  @Test
+  @DisplayName("PSPS - OK - recupero dei payments di un flow pubblicato")
+  void test_psp_getReportingFlowPaymentsPublished_Ok() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+
+    String url = GET_PAYMENTS_FDR_PUBLISHED_URL.formatted(PSP_CODE, flowName, 1L, EC_CODE);
+
+    PaginatedPaymentsResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(200)
+            .extract()
+            .as(PaginatedPaymentsResponse.class);
+    assertThat(res.getCount(), equalTo(5L));
+    List expectedList =
+        List.of(
+            PaymentStatusEnum.EXECUTED.name(),
+            PaymentStatusEnum.REVOKED.name(),
+            PaymentStatusEnum.NO_RPT.name(),
+            PaymentStatusEnum.STAND_IN.name(),
+            PaymentStatusEnum.STAND_IN_NO_RPT.name());
+    assertThat(
+        res.getData().stream().map(o -> o.getPayStatus().name()).toList(), equalTo(expectedList));
+    assertThat(
+        res.getData().stream().map(o -> o.getPayStatus().name()).toList(),
+        containsInAnyOrder(expectedList.toArray()));
+  }
+
+  @Test
+  @DisplayName("PSPS - OK - recupero dei flow created")
+  void test_psp_getReportingFlows_created_Ok() {
+    String flowName = TestUtil.getDynamicFlowName();
+    String urlPspFlow = FLOWS_URL.formatted(PSP_CODE, flowName);
+    String bodyFmtPspFlow =
+        FLOW_TEMPLATE.formatted(
+            flowName,
+            SenderTypeEnum.LEGAL_PERSON.name(),
+            PSP_CODE,
+            BROKER_CODE,
+            CHANNEL_CODE,
+            EC_CODE);
+
+    GenericResponse resPspFlow =
+        given()
+            .body(bodyFmtPspFlow)
+            .header(HEADER)
+            .when()
+            .post(urlPspFlow)
+            .then()
+            .statusCode(201)
+            .extract()
+            .body()
+            .as(GenericResponse.class);
+    assertThat(resPspFlow.getMessage(), equalTo(String.format("Fdr [%s] saved", flowName)));
+
+    String url = (GET_ALL_FDR_CREATED_URL).formatted(PSP_CODE);
+    PaginatedFlowsCreatedResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(200)
+            .extract()
+            .as(PaginatedFlowsCreatedResponse.class);
+
+    assertThat(res.getCount(), greaterThan(0L));
+    //    assertThat(res.getData(), contains(hasProperty("fdr", is(flowName))));
+    assertTrue(res.getData().stream().anyMatch(item -> item.getFdr().equals(flowName)));
+  }
+
+  @Test
+  @DisplayName("PSPS - OK - recupero dei payments di un flow creato")
+  void test_psp_getReportingFlowPayments_created_Ok() {
+    String flowName = TestUtil.getDynamicFlowName();
+    TestUtil.pspSunnyDay(flowName);
+    String url = (GET_PAYMENTS_FDR_CREATED_URL).formatted(PSP_CODE, flowName, EC_CODE);
+    PaginatedPaymentsResponse res =
+        given()
+            .header(HEADER)
+            .when()
+            .get(url)
+            .then()
+            .statusCode(200)
+            .extract()
+            .as(PaginatedPaymentsResponse.class);
+    List<Payment> data = res.getData();
+
+    assertThat(res.getCount(), equalTo(0L));
   }
 }
