+++ src/test/java/it/gov/pagopa/fdr/util/FileUtilTest.java
@@ -1,0 +1,60 @@
+package it.gov.pagopa.fdr.util;
+
+import io.quarkus.test.junit.QuarkusTest;
+import it.gov.pagopa.fdr.util.common.FileUtil;
+import it.gov.pagopa.fdr.util.error.enums.AppErrorCodeMessageEnum;
+import it.gov.pagopa.fdr.util.error.exception.common.AppException;
+import org.jboss.logging.Logger;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.*;
+
+@QuarkusTest
+class FileUtilTest {
+
+    private FileUtil fileUtil;
+    private Logger logger;
+
+    @BeforeEach
+    void setUp() {
+        logger = mock(Logger.class);
+        fileUtil = new FileUtil(logger);
+    }
+
+    @Test
+    @DisplayName("FileUtil getFileFromResourceAsStream test with existent file")
+    void testGetFileFromResourceAsStream_FileExists() {
+        InputStream result = fileUtil.getFileFromResourceAsStream("test.json");
+        assertNotNull(result);
+    }
+
+    @Test
+    @DisplayName("FileUtil getFileFromResourceAsStream test with non existent file")
+    void testGetFileFromResourceAsStream_FileNotFound() {
+        String fileName = "nonexistent.json";
+
+        AppException exception = assertThrows(AppException.class, () -> {
+            fileUtil.getFileFromResourceAsStream(fileName);
+        });
+
+        assertEquals(AppErrorCodeMessageEnum.FILE_UTILS_FILE_NOT_FOUND, exception.getCodeMessage());
+        verify(logger).errorf("Error reading file: [%s]", fileName);
+    }
+
+
+    @Test
+    @DisplayName("FileUtil convertToString test with existent content")
+    void testConvertToString_Success() {
+        String content = "test content";
+        InputStream inputStream = new ByteArrayInputStream(content.getBytes());
+
+        String result = fileUtil.convertToString(inputStream);
+
+        assertEquals(content, result);
+    }
+
+}
