+++ src/test/resources/test.json
@@ -1,0 +1,5 @@
+{
+  "test": "test",
+  "test2": "test2"
+
+}
